refactor: move libsignal to submodule
