+++ wacore/Cargo.toml
@@ -14,33 +14,19 @@ aes-gcm = { version = "0.10.2", features = [
     "alloc",
 ], default-features = false }
 anyhow = { version = "1.0", default-features = false }
-arrayref = "0.3.9"
 async-trait = "0.1.88"
 base64 = { version = "0.22.1", default-features = false, features = ["alloc"] }
 bytes = { version = "1.5", default-features = false }
-cbc = { version = "0.1.2", features = ["alloc"] }
 
 # Time handling
 chrono = { version = "0.4", features = [
     "now",
     "serde",
 ], default-features = false }
-ctr = "0.9.2"
-curve25519-dalek = { version = "4.1.3", features = ["digest"] }
-derive_more = { version = "2.0.1", features = [
-    "deref",
-    "from",
-    "into",
-    "try_from",
-] }
-displaydoc = "0.2.5"
 flate2 = { workspace = true, features = ["zlib-rs"], default-features = false }
-
-ghash = "0.5.1"
 hex = { version = "0.4", features = ["alloc"], default-features = false }
 hkdf = "0.12.3"
 hmac = { version = "0.12.1", features = ["reset"] }
-itertools = "0.14.0"
 log = "0.4.28"
 md5 = "0.8.0"
 once_cell = { version = "1.19", default-features = false, features = ["std"] }
@@ -51,12 +37,9 @@ rand = { version = "0.9" }
 rand_core = "0.9"
 serde = { version = "1.0.228", features = ["derive"] }
 serde-big-array = "0.5"
-sha1 = "0.10.6"
 sha2 = "0.10.6"
-subtle = "2.6.1"
 thiserror = { workspace = true }
 tokio = { version = "1.47.1", default-features = false, features = ["sync"] }
-uuid = { version = "1.18.1", default-features = false, features = ["v4"] }
 wacore-binary = { path = "./binary", version = "0.1.0", features = ["serde"] }
+wacore-libsignal = { path = "./libsignal", version = "0.1.0" }
 waproto = { path = "../waproto", version = "0.1.0" }
-x25519-dalek = { version = "2.0.1", features = ["static_secrets"] }
