+++ wacore/binary/Cargo.toml
