+++ wacore/libsignal/Cargo.toml
@@ -1,0 +1,47 @@
+[package]
+name = "wacore-libsignal"
+version = "0.1.0"
+edition = "2024"
+authors = ["João Lucas <jlucaso@hotmail.com>"]
+license = "MIT"
+repository = "https://github.com/jlucaso1/whatsapp-rust"
+description = "Signal Protocol implementation for the WhatsApp platform"
+
+[dependencies]
+aes = "0.8.3"
+aes-gcm = { version = "0.10.2", features = [
+  "aes",
+  "alloc",
+], default-features = false }
+arrayref = "0.3.9"
+async-trait = "0.1.88"
+cbc = { version = "0.1.2", features = ["alloc"] }
+chrono = { version = "0.4", features = [
+  "now",
+  "serde",
+], default-features = false }
+ctr = "0.9.2"
+curve25519-dalek = { version = "4.1.3", features = ["digest"] }
+derive_more = { version = "2.0.1", features = [
+  "deref",
+  "from",
+  "into",
+  "try_from",
+] }
+displaydoc = "0.2.5"
+ghash = "0.5.1"
+hex = { version = "0.4", features = ["alloc"], default-features = false }
+hkdf = "0.12.3"
+hmac = { version = "0.12.1", features = ["reset"] }
+itertools = "0.14.0"
+log = "0.4.28"
+prost = { version = "0.14.1", default-features = false }
+rand = { workspace = true }
+serde = { version = "1.0.228", features = ["derive"] }
+sha1 = "0.10.6"
+sha2 = "0.10.6"
+subtle = "2.6.1"
+thiserror = { workspace = true }
+uuid = { version = "1.18.1", default-features = false, features = ["v4"] }
+waproto = { path = "../../waproto", version = "0.1.0" }
+x25519-dalek = { version = "2.0.1", features = ["static_secrets"] }
