+++ wacore/libsignal/src/protocol/state/prekey.rs
@@ -7,7 +7,7 @@ use std::fmt;
 
 use prost::Message;
 
-use crate::libsignal::protocol::{
+use crate::protocol::{
     KeyPair, PrivateKey, PublicKey, Result, SignalProtocolError, stores::PreKeyRecordStructure,
 };
